+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ==========
 
+* Add `instance Serialize a => Serialize (NonEmpty a)`
 * Add `instance Serialize Void`
 
 0.5.8.1